(PC-23006)[PRO] refactor: ico-trash-filled to fullTrash and replace ico-ticket-plus-full with fullCode
