+++ pro/src/components/StockFormActions/types.ts
@@ -1,12 +1,6 @@
-import { FunctionComponent, SVGProps } from 'react'
-
 export interface StockFormRowAction {
   callback: () => void
   label: string
-  Icon?: FunctionComponent<
-    SVGProps<SVGSVGElement> & {
-      title?: string | undefined
-    }
-  >
+  icon?: string
   disabled?: boolean
 }
